User account page added

+++ src/main/java/com/example/crosswordgenerator/services/CrosswordService.java
@@ -4,9 +4,13 @@ import com.example.crosswordgenerator.models.Crossword;
 import com.example.crosswordgenerator.models.User;
 import com.example.crosswordgenerator.repositories.ICrosswordRepository;
 import com.example.crosswordgenerator.repositories.IUserRepository;
+
+import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.security.Principal;
 import java.util.List;
@@ -16,27 +20,31 @@ import java.util.List;
  * */
 @Service
 @Slf4j
-@RequiredArgsConstructor
 public class CrosswordService {
 
     private final ICrosswordRepository crosswordRepository;
     private final IUserRepository userRepository;
+    private final EntityManager entityManager;
+
+    public CrosswordService(ICrosswordRepository crosswordRepository, IUserRepository userRepository, EntityManager entitymanager) {
+        this.crosswordRepository = crosswordRepository;
+        this.userRepository = userRepository;
+        this.entityManager = entitymanager;
+    }
 
     /**
      * Получить все кроссворды.
      * @return список всех кроссвордов.
      * */
-    public List<Crossword> getAll() {
+    public Iterable<Crossword> getAll() {
         return crosswordRepository.findAll();
     }
 
     /**
      * Сохранить кроссворд.
-     * @param principal текущий зарегистрированный пользователь.
      * @param crossword кроссворд
      * */
-    public void save(Principal principal, Crossword crossword) {
-        crossword.setUser(getUserByPrincipal(principal));
+    public void save(Crossword crossword) {
         log.info("Saving new Crossword. Title: {}", crossword.getTitle());
         crosswordRepository.save(crossword);
     }
@@ -50,14 +58,21 @@ public class CrosswordService {
         if (principal == null) {
             return new User();
         }
-        return userRepository.findByUsername(principal.getName());
+        return userRepository.findByUsername(principal.getName()).orElse(null);
     }
     /**
      * Удаолить кроссворд.
-     * @param id id rроссворда, который должен быть удален.
+     * @param id id кроссворда, который должен быть удален.
      * */
-    public void delete(Long id) {
-        crosswordRepository.deleteById(id);
+    public boolean delete(Long id) {
+        if(crosswordRepository.findById(id).isPresent()){
+            log.info("Crossword with id {} deleted.", id);
+            crosswordRepository.deleteById(id);
+            return true;
+        }else{
+            log.info("Crossword with id {} not deleted because it doesn't exist", id);
+            return false;
+        }
     }
     /**
      * Возвращает кроссворд из базы данных по id.
@@ -66,5 +81,22 @@ public class CrosswordService {
      * */
     public Crossword getById(Long id) {
         return crosswordRepository.findById(id).orElse(null);
+    }
+
+    public boolean update(Crossword crossword){
+        if(crosswordRepository.findById(crossword.getId()).isPresent()){
+            crosswordRepository.save(crossword);
+            return true;
+        }else return false;
+    }
+
+    @Transactional
+    public boolean increaseSolvedCount(Long id){
+        Crossword cw = crosswordRepository.findById(id).orElse(null);
+        if(cw != null){
+            cw.setSolved(cw.getSolved() + 1);
+            crosswordRepository.save(cw);
+            return true;
+        }else return false;
     }
 }
